Closing contact links, updating club disband notice
Hey Jireh, 

If you are reading this, sure has been a while huh. 

Thanks for forming the club and setting up this website! You can take it down if you want, though I think it'll also be nice to keep it up so future NTU students can still read the old logs. 

Best of luck with whatever you are up to now :)

Warm regards,
Mingyue